split ssl usage
